modif insert

+++ src/ObjetBD.java
@@ -13,10 +13,10 @@ public class ObjetBD {
     }
     public void insereObjet(Objet o,Utilisateur u, Categorie cat){
         PreparedStatement s = this.laConnexionMySQL.preparedStatement("INSERT INTO OBJET VALUES (?,?,?,?,?)");
-        s.setDouble(1, o.getidObjet());
+        s.setInt(1, o.getidObjet());
         s.setString(2, o.getNomObjet());
         s.setString(3, o.getDescription());
-        s.setDouble(4, u.getIdUtilisateur());  
+        s.setInt(4, u.getIdUtilisateur());  
         s.setDouble(5, cat);
         s.executeQuery();
     }
